refac: small refactorings
