+++ Markdown2Pdf/Services/InlineOptionsParser.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using Markdown2Pdf.Options;
@@ -8,18 +9,16 @@ using YamlDotNet.Serialization.NamingConventions;
 namespace Markdown2Pdf.Services;
 internal class InlineOptionsParser {
 
-  public static bool TryParseYamlFrontMatter(string markdownFilePath, out Markdown2PdfOptions options) {
-    if (!_TryReadYamlFrontMatter(markdownFilePath, out var markdownContent)) {
-      options = null!;
-      return false;
-    }
+  public static Markdown2PdfOptions ParseYamlFrontMatter(string markdownFilePath) {
+    if (!_TryReadYamlFrontMatter(markdownFilePath, out var yamlContent))
+      throw new Exception($"Could not find a YAML front matter block in '{markdownFilePath}'.");
 
     var deserializer = new DeserializerBuilder()
       .WithNamingConvention(HyphenatedNamingConvention.Instance)
       .Build();
-    var yamlObject = deserializer.Deserialize<SerializableOptions>(markdownContent);
-    options = yamlObject.ToMarkdown2PdfOptions();
-    return true;
+
+    var options = deserializer.Deserialize<SerializableOptions>(yamlContent);
+    return options.ToMarkdown2PdfOptions();
   }
 
   private static bool _TryReadYamlFrontMatter(string markdownFilePath, out string markdownContent) {
@@ -51,4 +50,5 @@ internal class InlineOptionsParser {
     markdownContent = null!;
     return false;
   }
+
 }
